Agentic_Job_Search-main/pyproject.toml
[project]
name = "job-search-agent"
version = "0.1.0"
description = "AI-powered job search agent with memory using LangChain and ExaAI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-nebius>=0.1.3",
    "langchain-core>=0.1.0",
    "exa-py>=1.0.0",
    "memorisdk>=2.3.2",
    "streamlit>=1.50.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pypdf>=3.0.0",
]
